refactor: yosugaClientからyosugaEventを分離
