+++ src/index.ts
@@ -25,7 +25,7 @@ logger.debug(generateDependencyReport());
 
 //======================================================================
 
-export const client: Client = new Discord.Client({
+const client: Client = new Discord.Client({
   intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILD_VOICE_STATES],
 });
